add binary crates for full- and light-nodes
the full-node should eventually be finding hashes to advance a
blockchain, while also communicating and propagating information
(blocks, transactions) to other nodes. not sure if we also want to
have the full-node send transactions into the system (see light-node).

the light-node shouldn't be doing any heavy computation. instead it
should be used to send transactions into the network.

this is only a rough sketch though. everything is subject to change
at this point in time.