refactor (see notes)
moved the communication logic out of the node module. the communication
between the main node and it's communication part it a bit intertwined
but good enough for now. the same should happen with a future mining
module.
also "node" was renamed to "peer" in some places to make the code base
more consistent.
a basic killtoken was also added to be used with ctrlc's crtl-c handler
for graceful shutdowns.

there are some problems with dead locks that happen occasionally, needs
more investigating. maybe it's also just wrong timing on condvar notifies
(inside the Communication struct)